reorganize board controllers into package
