+++ controllers/board/__init__.py
@@ -1,0 +1,4 @@
+from .command_dispatcher import BoardCommandDispatcher
+from .controller import BoardController
+
+__all__ = ["BoardCommandDispatcher", "BoardController"]
