+++ controllers/board/controller.py
@@ -23,19 +23,19 @@ from core.board_edit.session import EditSessionState
 from core.board_apply_runtime import BoardApplyRuntime
 from core.board_io import backup_board_payload, board_path, load_board_payload, save_board_payload
 from core.board_media_cache import BoardMediaCache
-from controllers.board_edit_focus_controller import BoardEditFocusController
-from controllers.board_edit_panel_controller import BoardEditPanelController
-from controllers.board_edit_preview_controller import BoardEditPreviewController
-from controllers.board_edit_timeline_controller import BoardEditTimelineController
-from controllers.board_edit_tools_controller import BoardEditToolsController
-from controllers.board_group_actions_controller import BoardGroupActionsController
-from controllers.board_groups_controller import BoardGroupsController
-from controllers.board_history_controller import BoardHistoryController
-from controllers.board_legacy_payload_controller import BoardLegacyPayloadController
-from controllers.board_media_import_controller import BoardMediaImportController
-from controllers.board_media_render_controller import BoardMediaRenderController
-from controllers.board_notes_controller import BoardNotesController
-from controllers.board_scene_view_controller import BoardSceneViewController
+from controllers.board.edit_focus_controller import BoardEditFocusController
+from controllers.board.edit_panel_controller import BoardEditPanelController
+from controllers.board.edit_preview_controller import BoardEditPreviewController
+from controllers.board.edit_timeline_controller import BoardEditTimelineController
+from controllers.board.edit_tools_controller import BoardEditToolsController
+from controllers.board.group_actions_controller import BoardGroupActionsController
+from controllers.board.groups_controller import BoardGroupsController
+from controllers.board.history_controller import BoardHistoryController
+from controllers.board.legacy_payload_controller import BoardLegacyPayloadController
+from controllers.board.media_import_controller import BoardMediaImportController
+from controllers.board.media_render_controller import BoardMediaRenderController
+from controllers.board.notes_controller import BoardNotesController
+from controllers.board.scene_view_controller import BoardSceneViewController
 from core.board_state import (
     ApplyPayloadState,
     apply_pending_groups_to_scene,
