+++ controllers/board/edit_panel_controller.py
